move bagit profiles to separate app
